initial revamp for new item status.
