+++ frontend/src/atoms/statusAtoms.ts
@@ -1,0 +1,6 @@
+import { atom } from 'jotai';
+
+export const isFailedAtom = atom(false);
+export const isSuccessAtom = atom(false);
+export const errorDetailsAtom = atom<string | null>(null);
+export const errorMessageAtom = atom<string | null>(null);
